☀️ 73% – adopt the battle command action executor
main:0808e680 (748 bytes, BattleCommand_ExecuteSelectedAction) compiles
byte-exact once the battle work fields are reached through the shared
struct instead of absolute addresses, the action definition is fetched a
second time for the cost, and the 0x920 message id is a symbol so it stays
a pool word.

The function lives in translation unit unit-0808e23c, whose members must
agree on the event lookup's prototype. The executor needs the group
parameter declared s32 (a u16 parameter adds a truncation), so the unit's
declarations move to s32 and the one exact caller that relied on the
truncation spells it as a cast; both stay byte-exact.

+++ games/gs1/recon/en/main/0808e4b4.c
@@ -38,7 +38,7 @@ struct BattleEffectValueRecord {
  * narrower C type.
  */
 s32 BattleEffect_FindMatchingEvent(
-    s32 requested_flags, u16 group, void *result)
+    s32 requested_flags, s32 group, void *result)
 {
     struct BattleEffectRuntime *runtime =
         *(struct BattleEffectRuntime **)0x03001ebc;
